fix: address clang-tidy warnings across entire codebase
Fix HeaderFilterRegex in .clang-tidy to match endo project directories
instead of the old contour project paths.

Apply automated fixes for: designated-initializers, qualified-auto,
redundant-member-init, container-data-pointer, use-auto, use-emplace,
faster-string-find, move-const-arg, container-size-empty,
redundant-casting, use-scoped-lock, raw-string-literal,
use-std-min-max, loop-convert, integer-sign-comparison,
use-std-print, and use-ranges.

Manual fixes for: enum-size (uint8_t base types), narrowing-conversions
(explicit static_cast), too-small-loop-variable, empty-catch,
suspicious-stringview-data-usage, unused-return-value,
convert-member-functions-to-static, math-missing-parentheses,
make-member-function-const, inconsistent-declaration-parameter-name,
avoid-nested-conditional-operator, isolate-declaration,
deadcode.DeadStores, performance.Padding, pro-type-member-init,
init-variables, use-default-member-init, unnecessary-value-param,
and inefficient-vector-operation.

Also fixes a bug in Runner.cpp PCMPEQ handler that was comparing
getIPAddress(-2) == getIPAddress(-2) (self-comparison, always true)
instead of getIPAddress(-2) == getIPAddress(-1).

+++ src/agent/auth/OAuthCallbackServer.cpp
@@ -113,7 +113,7 @@ auto OAuthCallbackServer::start() -> std::expected<uint16_t, std::string>
         return std::unexpected(std::string("WSAStartup failed"));
 #endif
 
-    _listenFd = static_cast<int>(socket(AF_INET, SOCK_STREAM, 0));
+    _listenFd = socket(AF_INET, SOCK_STREAM, 0);
     if (_listenFd < 0)
         return std::unexpected(std::string("Failed to create socket: ") + strerror(errno));
 
@@ -188,7 +188,7 @@ auto OAuthCallbackServer::waitForCallback(std::chrono::seconds timeout)
 #endif
 
     // Accept the connection.
-    auto const clientFd = static_cast<int>(accept(_listenFd, nullptr, nullptr));
+    auto const clientFd = accept(_listenFd, nullptr, nullptr);
     if (clientFd < 0)
         return std::unexpected(std::string("Failed to accept connection: ") + strerror(errno));
